refactor(slack-runtime): extract api client and state store modules (#1292)
