+++ crates/tau-slack-runtime/src/slack_runtime.rs
@@ -1,14 +1,12 @@
 use std::{
-    collections::{HashMap, HashSet, VecDeque},
-    io::Write,
+    collections::{HashMap, VecDeque},
     path::{Path, PathBuf},
     sync::{Arc, Mutex},
     time::{Duration, Instant},
 };
 
-use anyhow::{anyhow, bail, Context, Result};
+use anyhow::{anyhow, Context, Result};
 use futures_util::{SinkExt, StreamExt};
-use serde::de::DeserializeOwned;
 use serde::{Deserialize, Serialize};
 use serde_json::{json, Value};
 use tau_agent_core::{Agent, AgentConfig, AgentEvent};
@@ -63,489 +61,11 @@ pub struct SlackBridgeRuntimeConfig {
     pub artifact_retention_days: u64,
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
-struct SlackBridgeState {
-    schema_version: u32,
-    #[serde(default)]
-    processed_event_keys: Vec<String>,
-    #[serde(default)]
-    health: TransportHealthSnapshot,
-}
+mod slack_api_client;
+mod slack_state_store;
 
-impl Default for SlackBridgeState {
-    fn default() -> Self {
-        Self {
-            schema_version: SLACK_STATE_SCHEMA_VERSION,
-            processed_event_keys: Vec::new(),
-            health: TransportHealthSnapshot::default(),
-        }
-    }
-}
-
-struct SlackBridgeStateStore {
-    path: PathBuf,
-    cap: usize,
-    state: SlackBridgeState,
-    processed_index: HashSet<String>,
-}
-
-impl SlackBridgeStateStore {
-    fn load(path: PathBuf, cap: usize) -> Result<Self> {
-        let mut state = if path.exists() {
-            let raw = std::fs::read_to_string(&path)
-                .with_context(|| format!("failed to read state file {}", path.display()))?;
-            serde_json::from_str::<SlackBridgeState>(&raw).with_context(|| {
-                format!("failed to parse slack bridge state file {}", path.display())
-            })?
-        } else {
-            SlackBridgeState::default()
-        };
-
-        if state.schema_version != SLACK_STATE_SCHEMA_VERSION {
-            bail!(
-                "unsupported slack bridge state schema: expected {}, found {}",
-                SLACK_STATE_SCHEMA_VERSION,
-                state.schema_version
-            );
-        }
-
-        let cap = cap.max(1);
-        if state.processed_event_keys.len() > cap {
-            let keep_from = state.processed_event_keys.len() - cap;
-            state.processed_event_keys = state.processed_event_keys[keep_from..].to_vec();
-        }
-
-        let processed_index = state
-            .processed_event_keys
-            .iter()
-            .cloned()
-            .collect::<HashSet<_>>();
-        Ok(Self {
-            path,
-            cap,
-            state,
-            processed_index,
-        })
-    }
-
-    fn contains(&self, key: &str) -> bool {
-        self.processed_index.contains(key)
-    }
-
-    fn mark_processed(&mut self, key: &str) -> bool {
-        if self.processed_index.contains(key) {
-            return false;
-        }
-        self.state.processed_event_keys.push(key.to_string());
-        self.processed_index.insert(key.to_string());
-        while self.state.processed_event_keys.len() > self.cap {
-            let removed = self.state.processed_event_keys.remove(0);
-            self.processed_index.remove(&removed);
-        }
-        true
-    }
-
-    fn transport_health(&self) -> &TransportHealthSnapshot {
-        &self.state.health
-    }
-
-    fn update_transport_health(&mut self, value: TransportHealthSnapshot) -> bool {
-        if self.state.health == value {
-            return false;
-        }
-        self.state.health = value;
-        true
-    }
-
-    fn save(&self) -> Result<()> {
-        let mut payload =
-            serde_json::to_string_pretty(&self.state).context("failed to serialize state")?;
-        payload.push('\n');
-        write_text_atomic(&self.path, &payload)
-            .with_context(|| format!("failed to write state file {}", self.path.display()))?;
-        Ok(())
-    }
-}
-
-#[derive(Clone)]
-struct JsonlEventLog {
-    path: PathBuf,
-    file: Arc<Mutex<std::fs::File>>,
-}
-
-impl JsonlEventLog {
-    fn open(path: PathBuf) -> Result<Self> {
-        if let Some(parent) = path.parent() {
-            if !parent.as_os_str().is_empty() {
-                std::fs::create_dir_all(parent)
-                    .with_context(|| format!("failed to create {}", parent.display()))?;
-            }
-        }
-        let file = std::fs::OpenOptions::new()
-            .create(true)
-            .append(true)
-            .open(&path)
-            .with_context(|| format!("failed to open {}", path.display()))?;
-        Ok(Self {
-            path,
-            file: Arc::new(Mutex::new(file)),
-        })
-    }
-
-    fn append(&self, value: &Value) -> Result<()> {
-        let line = serde_json::to_string(value).context("failed to encode log event")?;
-        let mut file = self
-            .file
-            .lock()
-            .map_err(|_| anyhow!("event log mutex is poisoned"))?;
-        writeln!(file, "{line}")
-            .with_context(|| format!("failed to append to {}", self.path.display()))?;
-        file.flush()
-            .with_context(|| format!("failed to flush {}", self.path.display()))?;
-        Ok(())
-    }
-}
-
-#[derive(Debug, Clone, Deserialize)]
-struct SlackAuthTestResponse {
-    ok: bool,
-    user_id: Option<String>,
-    error: Option<String>,
-}
-
-#[derive(Debug, Clone, Deserialize)]
-struct SlackOpenSocketResponse {
-    ok: bool,
-    url: Option<String>,
-    error: Option<String>,
-}
-
-#[derive(Debug, Clone, Deserialize)]
-struct SlackChatMessageResponse {
-    ok: bool,
-    ts: Option<String>,
-    channel: Option<String>,
-    error: Option<String>,
-}
-
-#[derive(Debug, Clone)]
-struct SlackPostedMessage {
-    channel: String,
-    ts: String,
-}
-
-#[derive(Clone)]
-struct SlackApiClient {
-    http: reqwest::Client,
-    api_base: String,
-    app_token: String,
-    bot_token: String,
-    retry_max_attempts: usize,
-    retry_base_delay_ms: u64,
-}
-
-impl SlackApiClient {
-    fn new(
-        api_base: String,
-        app_token: String,
-        bot_token: String,
-        request_timeout_ms: u64,
-        retry_max_attempts: usize,
-        retry_base_delay_ms: u64,
-    ) -> Result<Self> {
-        let mut headers = reqwest::header::HeaderMap::new();
-        headers.insert(
-            reqwest::header::USER_AGENT,
-            reqwest::header::HeaderValue::from_static("Tau-slack-bridge"),
-        );
-        headers.insert(
-            reqwest::header::ACCEPT,
-            reqwest::header::HeaderValue::from_static("application/json"),
-        );
-        let http = reqwest::Client::builder()
-            .default_headers(headers)
-            .timeout(Duration::from_millis(request_timeout_ms.max(1)))
-            .build()
-            .context("failed to create slack api client")?;
-
-        Ok(Self {
-            http,
-            api_base: api_base.trim_end_matches('/').to_string(),
-            app_token: app_token.trim().to_string(),
-            bot_token: bot_token.trim().to_string(),
-            retry_max_attempts: retry_max_attempts.max(1),
-            retry_base_delay_ms: retry_base_delay_ms.max(1),
-        })
-    }
-
-    async fn resolve_bot_user_id(&self) -> Result<String> {
-        let response: SlackAuthTestResponse = self
-            .request_json(
-                "auth.test",
-                || {
-                    self.http
-                        .post(format!("{}/auth.test", self.api_base))
-                        .bearer_auth(&self.bot_token)
-                },
-                true,
-            )
-            .await?;
-
-        if !response.ok {
-            bail!(
-                "slack auth.test failed: {}",
-                response
-                    .error
-                    .unwrap_or_else(|| "unknown error".to_string())
-            );
-        }
-
-        response
-            .user_id
-            .filter(|value| !value.trim().is_empty())
-            .ok_or_else(|| anyhow!("slack auth.test did not return user_id"))
-    }
-
-    async fn open_socket_connection(&self) -> Result<String> {
-        let response: SlackOpenSocketResponse = self
-            .request_json(
-                "apps.connections.open",
-                || {
-                    self.http
-                        .post(format!("{}/apps.connections.open", self.api_base))
-                        .bearer_auth(&self.app_token)
-                },
-                true,
-            )
-            .await?;
-        if !response.ok {
-            bail!(
-                "slack apps.connections.open failed: {}",
-                response
-                    .error
-                    .unwrap_or_else(|| "unknown error".to_string())
-            );
-        }
-        response
-            .url
-            .filter(|value| !value.trim().is_empty())
-            .ok_or_else(|| anyhow!("slack apps.connections.open did not return url"))
-    }
-
-    async fn post_message(
-        &self,
-        channel: &str,
-        text: &str,
-        thread_ts: Option<&str>,
-    ) -> Result<SlackPostedMessage> {
-        let mut payload = json!({
-            "channel": channel,
-            "text": text,
-            "mrkdwn": false,
-            "unfurl_links": false,
-            "unfurl_media": false,
-        });
-        if let Some(thread_ts) = thread_ts {
-            payload["thread_ts"] = Value::String(thread_ts.to_string());
-        }
-
-        let response: SlackChatMessageResponse = self
-            .request_json(
-                "chat.postMessage",
-                || {
-                    self.http
-                        .post(format!("{}/chat.postMessage", self.api_base))
-                        .bearer_auth(&self.bot_token)
-                        .json(&payload)
-                },
-                true,
-            )
-            .await?;
-
-        if !response.ok {
-            bail!(
-                "slack chat.postMessage failed: {}",
-                response
-                    .error
-                    .unwrap_or_else(|| "unknown error".to_string())
-            );
-        }
-
-        Ok(SlackPostedMessage {
-            channel: response.channel.unwrap_or_else(|| channel.to_string()),
-            ts: response
-                .ts
-                .ok_or_else(|| anyhow!("slack chat.postMessage response missing ts"))?,
-        })
-    }
-
-    async fn update_message(
-        &self,
-        channel: &str,
-        ts: &str,
-        text: &str,
-    ) -> Result<SlackPostedMessage> {
-        let payload = json!({
-            "channel": channel,
-            "ts": ts,
-            "text": text,
-            "mrkdwn": false,
-        });
-        let response: SlackChatMessageResponse = self
-            .request_json(
-                "chat.update",
-                || {
-                    self.http
-                        .post(format!("{}/chat.update", self.api_base))
-                        .bearer_auth(&self.bot_token)
-                        .json(&payload)
-                },
-                true,
-            )
-            .await?;
-        if !response.ok {
-            bail!(
-                "slack chat.update failed: {}",
-                response
-                    .error
-                    .unwrap_or_else(|| "unknown error".to_string())
-            );
-        }
-        Ok(SlackPostedMessage {
-            channel: response.channel.unwrap_or_else(|| channel.to_string()),
-            ts: response.ts.unwrap_or_else(|| ts.to_string()),
-        })
-    }
-
-    async fn download_file(&self, url: &str) -> Result<Vec<u8>> {
-        let request = || self.http.get(url).bearer_auth(&self.bot_token);
-        self.request_bytes("file download", request, false).await
-    }
-
-    async fn request_json<T, F>(
-        &self,
-        operation: &str,
-        mut builder: F,
-        decode_error_body: bool,
-    ) -> Result<T>
-    where
-        T: DeserializeOwned,
-        F: FnMut() -> reqwest::RequestBuilder,
-    {
-        let mut attempt = 0_usize;
-        loop {
-            attempt = attempt.saturating_add(1);
-            let response = builder()
-                .header("x-tau-retry-attempt", attempt.saturating_sub(1).to_string())
-                .send()
-                .await;
-            match response {
-                Ok(response) => {
-                    let status = response.status();
-                    if status.is_success() {
-                        let parsed = response
-                            .json::<T>()
-                            .await
-                            .with_context(|| format!("failed to decode slack {operation}"))?;
-                        return Ok(parsed);
-                    }
-
-                    let retry_after = parse_retry_after(response.headers());
-                    let body = if decode_error_body {
-                        response.text().await.unwrap_or_default()
-                    } else {
-                        String::new()
-                    };
-                    if attempt < self.retry_max_attempts
-                        && is_retryable_slack_status(status.as_u16())
-                    {
-                        tokio::time::sleep(retry_delay(
-                            self.retry_base_delay_ms,
-                            attempt,
-                            retry_after,
-                        ))
-                        .await;
-                        continue;
-                    }
-
-                    bail!(
-                        "slack api {operation} failed with status {}: {}",
-                        status.as_u16(),
-                        truncate_for_error(&body, 800)
-                    );
-                }
-                Err(error) => {
-                    if attempt < self.retry_max_attempts && is_retryable_transport_error(&error) {
-                        tokio::time::sleep(retry_delay(self.retry_base_delay_ms, attempt, None))
-                            .await;
-                        continue;
-                    }
-                    return Err(error)
-                        .with_context(|| format!("slack api {operation} request failed"));
-                }
-            }
-        }
-    }
-
-    async fn request_bytes<F>(
-        &self,
-        operation: &str,
-        mut builder: F,
-        decode_error_body: bool,
-    ) -> Result<Vec<u8>>
-    where
-        F: FnMut() -> reqwest::RequestBuilder,
-    {
-        let mut attempt = 0_usize;
-        loop {
-            attempt = attempt.saturating_add(1);
-            let response = builder()
-                .header("x-tau-retry-attempt", attempt.saturating_sub(1).to_string())
-                .send()
-                .await;
-            match response {
-                Ok(response) => {
-                    let status = response.status();
-                    if status.is_success() {
-                        return Ok(response.bytes().await?.to_vec());
-                    }
-                    let retry_after = parse_retry_after(response.headers());
-                    let body = if decode_error_body {
-                        response.text().await.unwrap_or_default()
-                    } else {
-                        String::new()
-                    };
-                    if attempt < self.retry_max_attempts
-                        && is_retryable_slack_status(status.as_u16())
-                    {
-                        tokio::time::sleep(retry_delay(
-                            self.retry_base_delay_ms,
-                            attempt,
-                            retry_after,
-                        ))
-                        .await;
-                        continue;
-                    }
-
-                    bail!(
-                        "slack api {operation} failed with status {}: {}",
-                        status.as_u16(),
-                        truncate_for_error(&body, 800)
-                    );
-                }
-                Err(error) => {
-                    if attempt < self.retry_max_attempts && is_retryable_transport_error(&error) {
-                        tokio::time::sleep(retry_delay(self.retry_base_delay_ms, attempt, None))
-                            .await;
-                        continue;
-                    }
-                    return Err(error)
-                        .with_context(|| format!("slack api {operation} request failed"));
-                }
-            }
-        }
-    }
-}
+use slack_api_client::{SlackApiClient, SlackPostedMessage};
+use slack_state_store::{JsonlEventLog, SlackBridgeStateStore};
 
 #[derive(Debug, Clone, Deserialize, Serialize)]
 struct SlackSocketEnvelope {
